requirements.txt
pygame==2.1.2
Pillow==9.0.0
pygame==2.1.2
Pillow==9.0.0
pygame==2.1.2
tk==0.1.0
Pillow==9.0.0
pygame==2.1.2
tk==0.1.0
